Clean up gcs tests directory and small fixes
