+++ CMakeLists.txt
@@ -41,7 +41,7 @@ option(VELOX_ENABLE_PARSE "Build parser used for unit tests." ON)
 option(VELOX_ENABLE_EXAMPLES
        "Build examples. This will enable VELOX_ENABLE_EXPRESSION automatically."
        OFF)
-option(VELOX_ENABLE_SUBSTRAIT "Buid Substrait-to-Velox converter." ON)
+option(VELOX_ENABLE_SUBSTRAIT "Buid Substrait-to-Velox converter." OFF)
 option(VELOX_ENABLE_BENCHMARKS "Build velox top level benchmarks." OFF)
 option(VELOX_ENABLE_BENCHMARKS_BASIC "Build velox basic benchmarks." OFF)
 option(VELOX_ENABLE_S3 "Build S3 Connector" OFF)
@@ -68,8 +68,6 @@ if(${VELOX_BUILD_MINIMAL})
   set(VELOX_ENABLE_SUBSTRAIT OFF)
   set(VELOX_CODEGEN_SUPPORT OFF)
 endif()
-set(VELOX_ENABLE_SUBSTRAIT ON)
-set(VELOX_ENABLE_GCS ON)
 
 if(${VELOX_BUILD_TESTING})
   # Enable all components to build testing binaries
@@ -126,7 +124,6 @@ if(VELOX_ENABLE_GCS)
   add_definitions(-DVELOX_ENABLE_GCS)
   if(GCSSDK_ROOT_DIR)
     set(CMAKE_PREFIX_PATH ${GCSSDK_ROOT_DIR})
-    message(STATUS "GCSDK root is ${GCSSDK_ROOT_DIR}!")
   else()
     message(FATAL_ERROR "no GCSSDK_ROOT_DIR provided!")
   endif()
