+++ velox/connectors/hive/storage_adapters/gcs/tests/GCSUtilTest.cpp
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "connectors/hive/storage_adapters/gcs/GCSUtil.h"
+
+#include "gtest/gtest.h"
+
+using namespace facebook::velox;
+
+TEST(GCSUtilTest, isGCSFile) {
+  EXPECT_FALSE(isGCSFile("gs:"));
+  EXPECT_FALSE(isGCSFile("gs::/bucket"));
+  EXPECT_FALSE(isGCSFile("gs:/bucket"));
+  EXPECT_TRUE(isGCSFile("gs://bucket/file.txt"));
+}
+
+
+TEST(GCSUtilTest, bucketAndKeyFromGCSPath) {
+  std::string bucket, key;
+  auto path = "bucket/file.txt";
+  bucketAndKeyFromGCSPath(path, bucket, key);
+  EXPECT_EQ(bucket, "bucket");
+  EXPECT_EQ(key, "file.txt");
+}
